+++ app/src/main/java/labs/zero_one/roundo/PreferenceActivity.kt
@@ -1,10 +1,11 @@
 package labs.zero_one.roundo
 
+import android.content.SharedPreferences
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.view.MenuItem
 import com.takisoft.fix.support.v7.preference.PreferenceFragmentCompat
-
+import org.jetbrains.anko.support.v4.defaultSharedPreferences
 
 /**
  * 主设置页面的 Activity
@@ -27,15 +28,47 @@ class PreferenceMainActivity : AppCompatActivity() {
      *
      * ## 重写方法列表
      * - [onCreatePreferencesFix]
+     * - [onDestroy]
+     * - [onSharedPreferenceChanged]
      *
      * @author lucka-me
      * @since 0.1.1
      */
-    class PreferenceMainFragment : PreferenceFragmentCompat() {
+    class PreferenceMainFragment :
+        PreferenceFragmentCompat(),
+        SharedPreferences.OnSharedPreferenceChangeListener {
 
         override fun onCreatePreferencesFix(savedInstanceState: Bundle?, rootKey: String?) {
             setPreferencesFromResource(R.xml.preference_main, rootKey)
 
+            if (defaultSharedPreferences
+                    .getBoolean(getString(R.string.pref_display_show_second_key), false)
+            ) {
+                findPreference(getString(R.string.pref_display_show_second_key)).icon =
+                    requireContext().getDrawable(R.drawable.ic_pref_display_show_second_ee)
+            }
+
+            defaultSharedPreferences.registerOnSharedPreferenceChangeListener(this)
+
+        }
+
+        override fun onDestroy() {
+            defaultSharedPreferences.unregisterOnSharedPreferenceChangeListener(this)
+            super.onDestroy()
+        }
+
+        override fun onSharedPreferenceChanged(sharedPreferences: SharedPreferences?, key: String?) {
+            if (key == null || sharedPreferences == null) return
+            if (key != getString(R.string.pref_display_show_second_key)) return
+            if (defaultSharedPreferences
+                    .getBoolean(getString(R.string.pref_display_show_second_key), false)
+            ) {
+                findPreference(getString(R.string.pref_display_show_second_key)).icon =
+                    requireContext().getDrawable(R.drawable.ic_pref_display_show_second_ee)
+            } else {
+                findPreference(getString(R.string.pref_display_show_second_key)).icon =
+                    requireContext().getDrawable(R.drawable.ic_pref_display_show_second)
+            }
         }
 
     }
@@ -101,11 +134,12 @@ class PreferenceAboutActivity : AppCompatActivity() {
                 context!!.packageManager.getPackageInfo(context!!.packageName, 0).versionName
             val versionCode =
                 context!!.packageManager.getPackageInfo(context!!.packageName, 0).versionCode
-            findPreference(getString(R.string.pref_about_summary_version_key)).summary = String.format(
-                getString(R.string.pref_about_summary_version_summary),
-                versionName,
-                versionCode
-            )
+            findPreference(getString(R.string.pref_about_summary_version_key)).summary =
+                String.format(
+                    getString(R.string.pref_about_summary_version_summary),
+                    versionName,
+                    versionCode
+                )
         }
 
     }